remove eat & speakbehavior files integrate roam interfaces set roam method in pachyderm set speak behavior in hippo adjust zookeeper to show desired format we still will need have an exterior main method, NOT run from zookeeper
