+++ src/Animal/Animal.java
@@ -1,6 +1,19 @@
 package Animal;
 
-public abstract class Animal implements SpeakBehavior, EatBehavior {
+public interface EatBehavior {
+    String eat();
+}
+
+public interface SpeakBehavior {
+    String makeNoise();
+}
+
+public interface RoamBehavior {
+    String roam();
+}
+
+public abstract class Animal implements SpeakBehavior, EatBehavior, RoamBehavior {
+    // I'd like these to be public so we don't have to run methods to retrieve them
     protected String name;
     protected String type;
 
@@ -22,10 +35,6 @@ public abstract class Animal implements SpeakBehavior, EatBehavior {
 
     public String sleep() {
         return String.format("%s is asleep!", this.name);
-    }
-
-    public String roam() {
-        return String.format("%s is roaming!", this.name);
     }
 
     public void printAnimal() {
