--- src/Animal/EatBehavior.java
@@ -1,5 +1,0 @@
-package Animal;
-
-public interface EatBehavior {
-    String eat();
-}
