--- src/Animal/SpeakBehavior.java
@@ -1,5 +1,0 @@
-package Animal;
-
-public interface SpeakBehavior {
-    String makeNoise();
-}
