+++ src/Pachyderm/Hippo.java
@@ -2,8 +2,8 @@ package Pachyderm;
 
 import Animal.SpeakBehavior;
 
-public class Hippo extends Pachyderm implements SpeakBehavior {
-
+public class Hippo extends Pachyderm implements SpeakBehavior{
+ 
     public Hippo(String name) {
         super(name);
     }
