+++ src/Pachyderm/Pachyderm.java
@@ -2,9 +2,13 @@ package Pachyderm;
 
 import Animal.*;
 
-public class Pachyderm extends Animal implements SpeakBehavior {
+abstract public class Pachyderm extends Animal implements RoamBehavior {
     public Pachyderm(String name) {
         super(name);
         this.type = "Pachyderm";
+    }
+
+    public String roam() {
+        return "STOMP. STOMP. STOMP";
     }
 }
